Created base object for UVData and Telescope
avoids a bunch of code duplication
