+++ uvdata/__init__.py
@@ -1,6 +1,7 @@
+from . import uvbase
+from . import parameter
 from . import uv
+from . import utils
 from . import uvfits
 from . import fhd
 from . import miriad
-from . import parameter
-from . import utils
